backend/requirements.txt
djangorestframework==3.12.4
Pillow==8.2.0
crispy-forms-gds==0.2.2